requirements.txt
google-adk>=0.5.0
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.2
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.28.0
redis>=5.0.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pytest>=7.4.0
black>=24.0.0
mypy>=1.8.0 